pyproject.toml
[tool.poetry]
name = "ils-middleware"
version = "0.1.0"
description = "Sinopia ILS Middleware using Apache Airflow"
authors = ["Jeremy Nelson <jpnelson@stanford.edu>", "Aaron Collier <aaron.collier@stanford.edu", "John Martin <suntzu@stanford.edu>"]
license = "Apache2"

[tool.poetry.dependencies]
python = "^3.8"
pymarc = "^4.1.1"
apache-airflow = "^2.1.4"
apache-airflow-providers-amazon = "^2.2.0"
SQLAlchemy = "^1.3.18"
typing-extensions = "^3.10.0"
rdflib = "^6.0.2"

[tool.poetry.dev-dependencies]
black = "21.8b0"
flake8 = "^3.9.2"
mypy = "^0.910"
pytest = "^6.2"
pytest-cov = "^2.12"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
